+++ test/flop_phoenix_test.exs
@@ -50,6 +50,8 @@ defmodule Flop.PhoenixTest do
       end)
       |> assign_new(:opts, fn -> [] end)
       |> assign_new(:target, fn -> nil end)
+      |> assign_new(:hide_age, fn -> false end)
+      |> assign_new(:show_age, fn -> true end)
 
     ~H"""
     <Flop.Phoenix.table
@@ -63,8 +65,11 @@ defmodule Flop.PhoenixTest do
     >
       <:col let={pet} label="Name" field={:name}><%= pet.name %></:col>
       <:col let={pet} label="Email" field={:email}><%= pet.email %></:col>
-      <:col let={pet} label="Age"><%= pet.age %></:col>
+      <:col let={pet} label="Age" hide={@hide_age} show={@show_age}>
+        <%= pet.age %>
+      </:col>
       <:col let={pet} label="Species" field={:species}><%= pet.species %></:col>
+      <:col>column without label</:col>
     </Flop.Phoenix.table>
     """
   end
@@ -921,9 +926,9 @@ defmodule Flop.PhoenixTest do
         )
 
       assert [_] = Floki.find(html, "tr.mungo")
-      assert [_, _, _, _] = Floki.find(html, "th.bean")
+      assert [_, _, _, _, _] = Floki.find(html, "th.bean")
       assert [_] = Floki.find(html, "tr.salt")
-      assert [_, _, _, _] = Floki.find(html, "td.tolerance")
+      assert [_, _, _, _, _] = Floki.find(html, "td.tolerance")
     end
 
     test "doesn't render table if items list is empty" do
@@ -934,6 +939,28 @@ defmodule Flop.PhoenixTest do
       html = render_table()
       assert [th] = Floki.find(html, "th:fl-contains('Age')")
       assert Floki.children(th, include_text: false) == []
+    end
+
+    test "conditionally hides a column" do
+      html = render_table()
+      assert [_] = Floki.find(html, "th:fl-contains('Age')")
+      assert [_] = Floki.find(html, "td:fl-contains('8')")
+
+      html = render_table(hide_age: false, show_age: true)
+      assert [_] = Floki.find(html, "th:fl-contains('Age')")
+      assert [_] = Floki.find(html, "td:fl-contains('8')")
+
+      html = render_table(hide_age: true, show_age: true)
+      assert [] = Floki.find(html, "th:fl-contains('Age')")
+      assert [] = Floki.find(html, "td:fl-contains('8')")
+
+      html = render_table(hide_age: false, show_age: false)
+      assert [] = Floki.find(html, "th:fl-contains('Age')")
+      assert [] = Floki.find(html, "td:fl-contains('8')")
+
+      html = render_table(hide_age: true, show_age: false)
+      assert [] = Floki.find(html, "th:fl-contains('Age')")
+      assert [] = Floki.find(html, "td:fl-contains('8')")
     end
 
     test "displays headers with sorting function" do
@@ -975,7 +1002,7 @@ defmodule Flop.PhoenixTest do
           }
         )
 
-      assert [th_name, th_email, th_age, th_species] = Floki.find(html, "th")
+      assert [th_name, th_email, th_age, th_species, _] = Floki.find(html, "th")
       assert Floki.attribute(th_name, "aria-sort") == []
       assert Floki.attribute(th_email, "aria-sort") == ["ascending"]
       assert Floki.attribute(th_age, "aria-sort") == []
@@ -991,7 +1018,7 @@ defmodule Flop.PhoenixTest do
           }
         )
 
-      assert [th_name, th_email, th_age, th_species] = Floki.find(html, "th")
+      assert [th_name, th_email, th_age, th_species, _] = Floki.find(html, "th")
       assert Floki.attribute(th_name, "aria-sort") == ["descending"]
       assert Floki.attribute(th_email, "aria-sort") == []
       assert Floki.attribute(th_age, "aria-sort") == []
@@ -1002,7 +1029,7 @@ defmodule Flop.PhoenixTest do
           meta: %Flop.Meta{flop: %Flop{order_by: [], order_directions: []}}
         )
 
-      assert [th_name, th_email, th_age, th_species] = Floki.find(html, "th")
+      assert [th_name, th_email, th_age, th_species, _] = Floki.find(html, "th")
       assert Floki.attribute(th_name, "aria-sort") == []
       assert Floki.attribute(th_email, "aria-sort") == []
       assert Floki.attribute(th_age, "aria-sort") == []